\name{counts1}
\alias{counts1}
\title{A matrix representing counts in the distinct values for the two samples}
\description{This is a K by 2 matrix, where K is the number of distinct values. It specifies the counts in the K distinct values for the two samples. The data is generated from two samples with mean shift.}
