\name{E2}
\alias{E2}
\title{An edge matrix representing a similarity graph}
\description{This is a matrix with the number of rows the number of edges in the similarity graph and 2 columns.  Each row records the subject indices of the two edges of in the similarity graph.  The subject indices of sample 1 is 1:100, and the subject indices of sample 2 is 101:250.}

