\name{getComdist}
\alias{getComdist}
\title{Get distance between two components}
\description{This function calculates the distance between two components.}
\usage{
getComdist(g1,g2,distance)
}
\arguments{
  \item{g1}{The distinct values in Component 1.}
  \item{g2}{The distinct values in Component 2.}
  \item{distance}{A K by K matrix, which is the distance matrix on the distinct values and K is the number of distinct values with at least one observation in either group.}
}
\seealso{
  \code{\link{getGraph}}
}

