\name{getR1R2}
\alias{getR1R2}
\title{Get intermediate results for g.tests function}
\description{This function calculates R1 and R2 quantities.}
\usage{
getR1R2(E, G1)
}
\arguments{
  \item{E}{A matrix with the number of rows the number of edges in the similarity graph and 2 columns.  Each row records the subject indices of the two ends of an edge in the similarity graph.}
  \item{G1}{The subject indices of sample 1.}
}
\seealso{
  \code{\link{g.tests}}
}

