\name{gTests}
\alias{gTests}
\title{Graph-Based Two-Sample Tests}
\description{This package includes four graph-based two-sample tests under the continuous setting and the discrete setting.}
\author{
Hao Chen and Jingru Zhang

Maintainer: Hao Chen (hxchen@ucdavis.edu)
}
\references{
Friedman J. and Rafsky L. (1979). Multivariate generalizations of the WaldWolfowitz
and Smirnov two-sample tests. The Annals of Statistics 7(4):697-717.

Chen, H. and Zhang, N. R. (2013). Graph-based tests for two-sample comparisons of categorical data. Statistica Sinica 23:1479-1503.

Chen, H. and Friedman, J. H. (2017). A new graph-based two-sample test for multivariate and object data. Journal of the American Statistical Association, 112:517, 397-409.

Chen, H., Chen, X. and Su, Y. (2017). A weighted edge-count two sample test for multivariate and object data. Journal of the American Statistical Association.

Zhang, J. and Chen, H. (2017). Graph-based two-sample tests for discrete data. arXiv:1711.04349
}

\seealso{
  \code{\link{g.tests}}
  \code{\link{g.tests_discrete}}
  \code{\link{getGraph}}
}
  

