\name{nnlink_Com}
\alias{nnlink_Com}
\title{Get components by nearest neighbor link algorithm}
\description{This function obtains components based on the nearest neighbor link algorithm.}
\usage{
nnlink_Com(distance)
}
\arguments{
  \item{distance}{A K by K matrix, which is the distance matrix on the distinct values and K is the number of distinct values with at least one observation in either group.}
}
\seealso{
  \code{\link{getGraph}}
}

