\name{gTests}
\alias{gTests}
\title{Graph-Based Two-Sample Tests}
\description{This package includes two tests for assessing whether multiple covariates in a treatment group and a matched control group are balanced in observational studies.
}
\author{
Hao Chen

Maintainer: Hao Chen (hxchen@ucdavis.edu)
}
\references{
Friedman J. and Rafsky L. Multivariate generalizations of the WaldWolfowitz
and Smirnov two-sample tests. The Annals of Statistics, 7(4):697-717, 1979.

Chen, H. and Friedman, J. H.  A new graph-based two-sample test for multivariate and object data. Journal of the American Statistical Association, 2016.

Chen, H., Chen, X. and Su, Y. A weighted edge-count two sample test for multivariate and object data. arXiv:1604.06515.
}

\seealso{
  \code{\link{g.tests}}
}
  

